+++ blocklets/ai-runtime/api/src/routes/auth/index.ts
@@ -1,0 +1,8 @@
+import { walletHandler } from '@api/libs/auth';
+import { Router } from 'express';
+
+import verifyVC from './verify-vc';
+
+export function attachWalletHandlers(router: Router) {
+  walletHandler.attach({ app: router, ...verifyVC });
+}