+++ blocklets/ai-studio/src/locales/en.ts
@@ -878,4 +878,10 @@ export default flatten({
   multiple: 'Multiple',
   dropdown: 'Dropdown',
   checkbox: 'Checkbox',
+  noPreviewImage: 'No Preview Image',
+  verifyVC: 'Verify VC',
+  vcItem: 'VC Name',
+  trustedIssuers: 'Trusted Issuers',
+  buttonTitle: 'Button Title',
+  buttonTitleVerified: 'Button Title (verified)',
 });