+++ blocklets/ai-studio/src/locales/zh.ts
@@ -869,4 +869,10 @@ export default flatten({
   multiple: '多选',
   dropdown: '下拉选择',
   checkbox: '平铺选择',
+  noPreviewImage: '暂无预览图',
+  verifyVC: '验证 VC',
+  vcItem: 'VC 名称',
+  trustedIssuers: '信任的发布者',
+  buttonTitle: '按钮标题',
+  buttonTitleVerified: '按钮标题(验证成功)',
 });