+++ blocklets/ai-runtime/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.4.108 (2024-10-27)
+
+- refactor: optimize decision function name generator
+
+## 0.4.107 (2024-10-27)
+
+- fix: auto login when running agent
+
+## 0.4.106 (2024-10-27)
+
+- feat: support component schema filter & preview
+
+## 0.4.105 (2024-10-26)
+
+- fix: all fields must be required for json schema mode
+
+## 0.4.104 (2024-10-25)
+
+- feat: add vc input support
+
 ## 0.4.103 (2024-10-25)
 
 - fix: support integer parameter for open api